docker/proxy: don't use -linkshared unconditionally
Source: meta-virtualization
MR: 123707
Type: Integration
Disposition: Merged from meta-virtualization
ChangeID: 2d44f8e5bc68778881ace29bdf74bd48c1be5fcd
Description:

The linkshared is not supported in some machines like riscv64 and
when supported we can use the GO_LINKSHARED instaed.
So export GO_LINKSHARED on the recipe to be available for Makefile.

This is currently only used in libnetwork for the proxy build, but
could be used in additional locations in the future.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>