docker: add mobyproject:moby to CVE_PRODUCT
Source: meta-virtualization
MR: 123707
Type: Integration
Disposition: Merged from meta-virtualization
ChangeID: b4b38e8e43ec2516793135f5ab8f3206c086ee53
Description:

Add mobyproject:moby to CVE_PRODUCT to reflect where the source
is coming from for both docker recipes.

We keep the old 'docker' designation for compatibility.

It is unclear whether or not we should also be adding the cli
and libnetwork to the CVE_PRODUCT. But since they are on
different SRCREVs and not vendored, we keep them out for now.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>