cockpit-machines: add recipe
`cockpit-machines` is a Cockpit plugin used to provide a web interface
to manage virtual machines. Since Cockpit version 242, cockpit-machines
has been moved to a dedicated repository. This package is dependent on
cockpit package from meta-webserver layer.

To test it, access the cockpit-machines UI on the tested target at
`https://127.0.0.1:9090/machines`.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>